+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 ### Ajouts
 - Ajout d'un lien vers http://dss.nowina.lu pour validation de signature #385
 - Validation des fichiers XML et des signatures PADES des actes pour le super admin 
+- Export de la liste des collectvités pour le super admin ou l'admin de groupe #300
 
 ### Retraits